Simplified global data and moved more loading to components
