+++ apps/admin-x-settings/src/components/settings/membership/Tiers.tsx
@@ -7,13 +7,15 @@ import TiersList from './tiers/TiersList';
 import useRouting from '../../../hooks/useRouting';
 import {Tier} from '../../../types/api';
 import {checkStripeEnabled, getActiveTiers, getArchivedTiers} from '../../../utils/helpers';
-import {useGlobalData} from '../../providers/DataProvider';
+import {useBrowseTiers} from '../../../utils/api/tiers';
+import {useGlobalData} from '../../providers/GlobalDataProvider';
 
 const Tiers: React.FC<{ keywords: string[] }> = ({keywords}) => {
     const [selectedTab, setSelectedTab] = useState('active-tiers');
-    const {tiers, settings, config} = useGlobalData();
-    const activeTiers = getActiveTiers(tiers);
-    const archivedTiers = getArchivedTiers(tiers);
+    const {settings, config} = useGlobalData();
+    const {data: {tiers} = {}} = useBrowseTiers();
+    const activeTiers = getActiveTiers(tiers || []);
+    const archivedTiers = getArchivedTiers(tiers || []);
     const {updateRoute} = useRouting();
 
     const openConnectModal = () => {
