+++ apps/admin-x-settings/src/components/settings/membership/portal/AccountPage.tsx
@@ -3,7 +3,7 @@ import React, {FocusEventHandler, useState} from 'react';
 import TextField from '../../../../admin-x-ds/global/form/TextField';
 import {Setting, SettingValue} from '../../../../types/api';
 import {fullEmailAddress, getEmailDomain, getSettingValues} from '../../../../utils/helpers';
-import {useGlobalData} from '../../../providers/DataProvider';
+import {useGlobalData} from '../../../providers/GlobalDataProvider';
 
 const AccountPage: React.FC<{
     localSettings: Setting[]
