+++ apps/admin-x-settings/src/hooks/useSettingGroup.tsx
@@ -1,9 +1,9 @@
-import React, { useEffect, useRef, useState } from 'react';
-import useForm, { SaveState } from './useForm';
+import React, {useEffect, useRef, useState} from 'react';
+import useForm, {SaveState} from './useForm';
 import useGlobalDirtyState from './useGlobalDirtyState';
-import { Setting, SettingValue, SiteData } from '../types/api';
-import { useEditSettings } from '../utils/api/settings';
-import { useGlobalData } from '../components/providers/DataProvider';
+import {Setting, SettingValue, SiteData} from '../types/api';
+import {useEditSettings} from '../utils/api/settings';
+import {useGlobalData} from '../components/providers/GlobalDataProvider';
 
 interface LocalSetting extends Setting {
     dirty?: boolean;
